requirements.txt
Django==5.1.4
django-extensions==3.2.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
whitenoise==6.8.2
autopep8==2.3.1
isort==5.13.2
requests==2.32.3
click==8.1.7
djhtml==3.0.7
six==1.17.0
django-widget-tweaks==1.5.0
django-compressor==4.5.1
django-relativedelta==2.0.0
python-dateutil==2.9.0.post0
pillow==11.0.0
django-seed==0.3.1
jupyter==1.1.1
Markdown==3.7
langchain-groq==0.2.2

